ARM: bcm2835: fix 64-bit build warning in mbox
Fixes:
arch/arm/mach-bcm283x/mbox.c: In function ‘bcm2835_mbox_call_prop’:
arch/arm/mach-bcm283x/mbox.c:118:48: warning: cast from pointer to integer of different size [-Wpointer-to-int-cast]
arch/arm/mach-bcm283x/mbox.c:126:29: warning: cast from pointer to integer of different size [-Wpointer-to-int-cast]

+++ arch/arm/mach-bcm283x/mbox.c
@@ -115,7 +115,9 @@ int bcm2835_mbox_call_prop(u32 chan, struct bcm2835_mbox_hdr *buffer)
 			   (unsigned long)((void *)buffer +
 			   roundup(buffer->buf_size, ARCH_DMA_MINALIGN)));
 
-	ret = bcm2835_mbox_call_raw(chan, phys_to_bus((u32)buffer), &rbuffer);
+	ret = bcm2835_mbox_call_raw(chan,
+				    phys_to_bus((unsigned long)buffer),
+				    &rbuffer);
 	if (ret)
 		return ret;
 
@@ -123,7 +125,7 @@ int bcm2835_mbox_call_prop(u32 chan, struct bcm2835_mbox_hdr *buffer)
 				(unsigned long)((void *)buffer +
 				roundup(buffer->buf_size, ARCH_DMA_MINALIGN)));
 
-	if (rbuffer != phys_to_bus((u32)buffer)) {
+	if (rbuffer != phys_to_bus((unsigned long)buffer)) {
 		printf("mbox: Response buffer mismatch\n");
 		return -1;
 	}
